GenieTalk {by Aadarsh Chauhan}
This project is a simple conversational AI chatbot built using GPT-2, a state-of-the-art language model from Hugging Face's Transformers library. The chatbot is capable of understanding and generating human-like text responses based on user input. It utilizes Flask to create an API that accepts user queries and returns generated responses.